WIP: plot all montages to see where are we standing
